+++ CHANGELOG.md
@@ -4,10 +4,12 @@
 - Features:
   - Renaming some roles' variables, to use the same terminology as the official app (only the features already previously coded work)
   - Adding more informations on the roles' reference sheet (script version, Bootlegger rule, Stormcaught character)
+  - Glassmorphism for modals & dialogs
 - Fixes:
   - Sounds during a vote have been changed to be less disturbing
   - Host refreshing the page during a game with custom character does not give a blank screen anymore (but still requires reloading the script. Additional fix needed to fix completely the issue)
   - Avoid raising hand when pressing space in an input field
+  - Better readability in streamer mode
 - Localization:
 
 - Maintenance: